+++ UserFlightSearchWindow.xaml.cs
@@ -71,7 +71,7 @@ namespace Software_Engineering_Project
                             TransactionObj newTransact = new TransactionObj();
                             newTransact.FlightUID = selected.flightID.ToString();
                             newTransact.UserUID = App.LoggedInUser.uniqueID;
-                            newTransact.transactionAmt = selected.ticketPrice;
+                            newTransact.transactionAmt = (float)selected.ticketPrice;
                             App.TransactionHist.Add(newTransact);
 
                         }
@@ -95,7 +95,7 @@ namespace Software_Engineering_Project
                         TransactionObj newTransact = new TransactionObj();
                         newTransact.FlightUID = selected.flightID.ToString();
                         newTransact.UserUID = App.LoggedInUser.uniqueID;
-                        newTransact.transactionAmt = selected.ticketPrice;
+                        newTransact.transactionAmt = (float)selected.ticketPrice;
                         App.TransactionHist.Add(newTransact);
                     }
                 }
